requirements.txt
click==6.7
Flask==0.12.2
hashlib==20081119
hmac==20101005
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
Werkzeug==0.12.2
